make sure python 3.10.13 is used instead of 3.10.12
