+++ .github/workflows/test-action.yml
@@ -24,7 +24,7 @@ jobs:
           - 3005.1
           - 3006.1
         include:
-          - python-version: 3.10.12
+          - python-version: 3.10.13
             salt-version: 3006.1
 
     steps:
@@ -285,7 +285,7 @@ jobs:
         include:
           - python-version: 3.9
             salt-version: 3005.1
-          - python-version: 3.10.12
+          - python-version: 3.10.13
             salt-version: 3006.0rc1
 
     steps:
